remove snapshot transactions, make snapshots immutable
